updating filtering, sensors, and observations

+++ src/sensors/sensors.jl
@@ -12,11 +12,11 @@ include("fovn.jl")
 include("fov3.jl")
 
 function observe(m::SearchDomain, s::Sensor, p::Pose)
-	error("`observe` not defined for this sensor.")
+    error("`observe` not defined for this sensor.")
 end
 
 # Required for particle filter
 # does not require probability, returns density
-function O(s::Sensor, theta::LocTuple, p::Pose, o::Float64)
-	error("`O` (density version) not defined for this sensor.")
+function O(s::Sensor, theta::LocTuple, p::Pose, o)
+    error("`O` (density version) not defined for this sensor.")
 end
